sms/project/requirements.txt
asgiref==3.7.2
Django==3.2.25
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.7.1
Pillow==9.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
pyjwt==2.8.0